Optimize group  interval settings and change test wait condition
