+++ cmd/alertmanager/testdata/alertmanager.test.yaml
@@ -26,7 +26,7 @@ route:
 
   # When the first notification was sent, wait 'group_interval' to send a batch
   # of new alerts that started firing for that group.
-  group_interval: 10s
+  group_interval: 45s
 
   # If an alert has successfully been sent, wait 'repeat_interval' to
   # resend them.
